% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BBSmap.R
\name{BBSmap}
\alias{BBSmap}
\title{Read in BBS Relative Abundance Shapefile}
\usage{
data(Birds)
}
\arguments{
\item{SpeciesCode}{= x - vector of species alpha codes}
}
\value{
returns a shapefile with relative abundance information;
if more than one alpha code is supplied returns a list of shapfiles named
with the alpha code
}
\description{
Read in BBS Relative Abundance Shapefile
}

