% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Birds.R
\docType{data}
\name{Birds}
\alias{Birds}
\title{Table to look up Alpha Code}
\format{\preformatted{'data.frame':	1045 obs. of  4 variables:
 $ Common_Name   : Factor w/ 1045 levels "(Audubon's Warbler) Yellow-rumped Warbler",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ AOU           : int  6560 1740 1691 5690 3380 1331 6550 5671 5680 4130 ...
 $ Scientifc_Name: Factor w/ 950 levels "Acanthis flammea",..: NA NA NA NA NA NA NA NA NA NA ...
 $ FourLetterCode: Factor w/ 963 levels "ABDU","ABTO",..: 956 154 777 287 710 519 554 288 289 564 ...
}}
\usage{
data(Birds)
}
\description{
Data from Institute for Bird Populations and USGS
http://www.birdpop.org/pages/birdSpeciesCodes.php
https://www.pwrc.usgs.gov/bbl/manual/speclist.cfm
}
\examples{
data(Birds)
head(Birds)
}
\keyword{datasets}

