% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getH2map.R
\name{getH2map}
\alias{getH2map}
\title{Mean Annual H2 in precipitation - raster}
\usage{
\code{getH2map(NorthAmerica=TRUE,Globe=FALSE)}
}
\arguments{
\item{\code{NorthAmerica}}{logical defaults to \code{TRUE},}

\item{\code{Globe}}{logical defaults to \code{FALSE}}
}
\value{
a raster containing the mean annual H2 in precipitation for North America, the Globe or a list containing both.
Both maps have a North America Albers Equal Area Conic projection.
}
\description{
Mean Annual H2 in precipitation - raster
}

