% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readBorders.R
\name{readBorders}
\alias{readBorders}
\title{Shapefiles for Country and State Borders}
\usage{
\code{readBorders(Countries=TRUE,States=FALSE)}
}
\arguments{
\item{\code{Countries}}{logical defaults to \code{TRUE},}

\item{\code{States}}{logical defaults to \code{FALSE}}
}
\value{
a shapefile of either world countries, United States with state boundaries or a list containing both.
Both maps have a North America Albers Equal Area Conic projection.
}
\description{
Shapefiles for Country and State Borders
}

