% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BBSmap.R
\name{BBSmap}
\alias{BBSmap}
\title{Read in BBS Relative Abundance Shapefile}
\usage{
\code{BBSmap(SpeciesCode="Species Alpha Code", MakeRaster=TRUE,RelAbun=FALSE,res=10000,SnapToH2=FALSE)}
}
\arguments{
\item{\code{SpeciesCode}}{= x - vector of species alpha codes,}

\item{\code{MakeRaster}}{defaults = \code{TRUE}}

\item{\code{RelAbun}}{default = \code{FALSE}}

\item{\code{SnapToH2}}{default = \code{FALSE}}
}
\value{
a shapefile or raster with relative abundance information;
if more than one alpha code is supplied returns a list of shapfiles named
with the alpha code
}
\description{
Read in BBS Relative Abundance Shapefile
}

